refactor: location Entity 형태 제거 및 LocationEntity를 모델로 빼서 리팩터링
