+++ data/src/main/java/com/wakeup/data/Mapper.kt
@@ -1,19 +1,17 @@
-package com.wakeup.data.source
+package com.wakeup.data
 
 import com.wakeup.data.database.entity.GlobeEntity
-import com.wakeup.data.database.entity.LocationEntity
 import com.wakeup.data.database.entity.MomentEntity
 import com.wakeup.data.database.entity.PictureEntity
 import com.wakeup.domain.model.Moment
 
 fun MomentEntity.toDomain(
-    location: LocationEntity,
     pictures: List<PictureEntity>,
     globes: List<GlobeEntity>
 ): Moment =
     Moment(
-        location.mainAddress,
-        location.detailAddress,
+        this.location.mainAddress,
+        this.location.detailAddress,
         pictures.map { it.bitmap },
         this.content,
         globes.map { it.name },
