+++ data/src/main/java/com/wakeup/data/database/dao/MomentDao.kt
@@ -3,7 +3,6 @@ package com.wakeup.data.database.dao
 import androidx.room.Dao
 import androidx.room.Query
 import com.wakeup.data.database.entity.GlobeEntity
-import com.wakeup.data.database.entity.LocationEntity
 import com.wakeup.data.database.entity.MomentEntity
 import com.wakeup.data.database.entity.PictureEntity
 
@@ -13,9 +12,6 @@ import kotlinx.coroutines.flow.Flow
 interface MomentDao {
     @Query("SELECT * FROM moment")
     fun getMoments(): Flow<List<MomentEntity>>
-
-    @Query("SELECT * FROM location WHERE id = :locationId")
-    fun getLocation(locationId: Int): Flow<LocationEntity>
 
     @Query(
         "SELECT * FROM picture WHERE id IN" +
@@ -28,5 +24,4 @@ interface MomentDao {
                 "(SELECT globe_id FROM moment_globe WHERE moment_id = :momentId)"
     )
     fun getGlobes(momentId: Int): Flow<List<GlobeEntity>>
-
 }
