+++ data/src/main/java/com/wakeup/data/database/entity/MomentEntity.kt
@@ -1,10 +1,12 @@
 package com.wakeup.data.database.entity
 
 import androidx.room.ColumnInfo
+import androidx.room.Embedded
 import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.ForeignKey.CASCADE
 import androidx.room.PrimaryKey
+import com.wakeup.data.model.LocationEntity
 
 @Entity(
     tableName = "moment",
@@ -25,7 +27,7 @@ import androidx.room.PrimaryKey
 )
 data class MomentEntity(
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
-    @ColumnInfo(name = "location_id", index = true) val locationId: Int,
+    @Embedded @ColumnInfo val location: LocationEntity,
     @ColumnInfo(name = "thumbnail_id", index = true) val thumbnailId: Int,
     @ColumnInfo(name = "content") val content: String,
     @ColumnInfo(name = "date") val date: String,
