+++ data/src/main/java/com/wakeup/data/model/LocationEntity.kt
@@ -1,0 +1,9 @@
+package com.wakeup.data.model
+
+data class LocationEntity(
+    val id: Int = 0,
+    val mainAddress: String,
+    val detailAddress: String,
+    val latitude: Double,
+    val longitude: Double,
+)
