+++ data/src/main/java/com/wakeup/data/source/local/moment/MomentLocalDataSourceImpl.kt
@@ -2,7 +2,7 @@ package com.wakeup.data.source.local.moment
 
 import com.wakeup.data.database.dao.MomentDao
 import com.wakeup.data.database.entity.GlobeEntity
-import com.wakeup.data.database.entity.LocationEntity
+import com.wakeup.data.model.LocationEntity
 import com.wakeup.data.database.entity.MomentEntity
 import com.wakeup.data.database.entity.PictureEntity
 import kotlinx.coroutines.flow.Flow
@@ -14,14 +14,10 @@ class MomentLocalDataSourceImpl(
     override fun getMoments(): Flow<List<MomentEntity>> =
         momentDao.getMoments()
 
-    override fun getLocation(momentId: Int): Flow<LocationEntity> =
-        momentDao.getLocation(momentId)
-
     override fun getPictures(momentId: Int): Flow<List<PictureEntity>> =
         momentDao.getPictures(momentId)
 
     override fun getGlobes(momentId: Int): Flow<List<GlobeEntity>> =
         momentDao.getGlobes(momentId)
-
 
 }
