+++ domain/src/main/java/com/wakeup/domain/model/Location.kt
@@ -1,0 +1,8 @@
+package com.wakeup.domain.model
+
+data class Location(
+    val mainAddress: String,
+    val detailAddress: String,
+    val latitude: Double,
+    val longitude: Double
+)
